Configure webpack to provide a dummy fs at runtime
Something we have imports fs at runtime despite not using it, but this small
Webpack config change solves that problem for all modules in the future.

It's worth noting other MDX projects don't require this, so we should
investigate why we need this sometime. I tried, but the stack traces aren't too
friendly...